[all]update version, [login]password and token

+++ src/main/java/server/GithubToken.java
@@ -5,12 +5,14 @@ import exception.RequestException;
 import model.Encryption;
 import org.kohsuke.github.GitHub;
 
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.Serializable;
 
 /**
  * @author kamontat
- * @version 1.0
+ * @version 2.2
  * @since 1/28/2017 AD - 12:51 AM
  */
 public class GithubToken implements Serializable {
@@ -66,14 +68,24 @@ public class GithubToken implements Serializable {
 	
 	public static GithubToken loadCache(String password) {
 		try {
-			return Cache.loadCache(FILE_NAME).loadFromFile(GithubToken.class).decryptGT(password);
+			if (Cache.loadCache(FILE_NAME).isExist())
+				return Cache.loadCache(FILE_NAME).loadFromFile(GithubToken.class).decryptGT(password);
+			else return GithubToken.getEmptyToken();
 		} catch (Exception e) {
 			return GithubToken.getEmptyToken();
 		}
 	}
 	
-	public void removeCache() {
-		Cache.loadCache(FILE_NAME).delete();
+	public static void removeCache() {
+		if (Cache.loadCache(FILE_NAME).isExist()) Cache.loadCache(FILE_NAME).delete();
+	}
+	
+	public boolean isEmptyToken() {
+		return token == null;
+	}
+	
+	public static boolean haveCache() {
+		return Cache.loadCache(FILE_NAME).isExist();
 	}
 	
 	public static String getHelp() {
